tmap to yaml export - adding meta info to nodes
When a yaml file is created from a tmap, it misses some tags and so is not as per the (yaml) format for topological map. So a yaml file exported from tmap cannot be imported to mongodb.

A small fix is done by adding some meta tag to the objects in the yaml file