max_vel_trans fix for melodic
max_trans_vel is replaced by max_vel_trans in melodic. As of now, this parameter is reconfigured after each edge in execute_policy_server.py, but not in navigate.py. To make this backward compatible with kinetic, both max_trans_vel and max_vel_trans are configured for setting. However, only the existing parameters will be tried to be reconfigured now.