Efficient computing of closest edges in localisation.
Option to get the closest edges only from the N closest nodes to the robot.
Useful for very large and dense maps (such as clockhouse vanity transportation map) where you do not want to be
computing the distance from every edge in the map to the robot.

N set by rosparam `/topological_localisation/NumClosestNodes`.
Default is 0, such that that the distance is computed for every edge in the map.