Auto set goal when adding an edge to the topological map
If a file e.g. `thorvald_navigation_actions/config/inrownavGoal.yaml` exists then the
map manager will automatically set the goal of the action to the goal specified in thaty file,
when the add edge function/service is called with arg `action_type` set to `inrownavGoal`.
If that arg is not set then a move base type goal is assumed.